[MIG] l10n_br_portal: Migration to 18.0
