+++ l10n_br_portal/views/portal_templates.xml
@@ -96,7 +96,7 @@
                                     type="text"
                                     name="vat"
                                     t-attf-class="input-vat form-control #{error.get('vat') and 'is-invalid' or ''}"
-                                    t-att-value="vat or partner.vat"
+                                    t-att-value="vat or partner.vat_formatted_cnpj"
                                     t-att-readonly="None if partner.can_edit_vat() else '1'"
                                     t-att-title="None if partner.can_edit_vat() else 'Changing CNPJ/CPF number is not allowed once document(s) have been issued for your account. Please contact us directly for this operation.'"
                                 />
